refactor: 로직 분리
